backend/trustgrid-api/requirements.txt
fastapi
uvicorn[standard]
pymongo
pydantic
pydantic-settings
google-generativeai
python-dotenv
passlib[bcrypt]